### 2019-01-03 #274 @pachamaltese
tradestatistics package
Submitting Author: Pachá (@pachamaltese)
Repository: https://github.com/tradestatistics/tradestatistics
Version submitted: 0.1
Version accepted: TBD

---

```
Package: tradestatistics
Title: Open Trade Statistics API Wrapper and Utility Program
Version: 0.1
  person("Mauricio", "Vargas S.", role = c("aut", "cre", "cph"), email = "mvargas@dcc.uchile.cl"),
  person("Joshua", "Kunst", role = c("ctb")),
  person("Amanda", "Dobbyn", role = c("ctb")),
  person("Jorge", "Cimentada", role = c("ctb")),
  person(family = "UN Comtrade", role = c("dtc"))
  )
Description: Access Open Trade Statistics API from R to download international trade data.
  crul,
  jsonlite
  covr,
  styler,
  pkgdown,
  vcr
```

- Explain how the and why the package falls under these categories (briefly, 1-2 sentences):

Because the package connects to an API and does API calls to simplify things for the final user who wants imports/exports and some metrics such as % of increase/decrease.

Non-expert users that use international trade data. This can also be targeted to intermediate/advanced users who can benefit from the speed and short syntax that this package provides.

No, this is made for a new data project

No.
