### 2019-01-04 #274 @pachamaltese
@maelle Hi!! Yes, I added all the feedback from the previous package and I extended it to obtain data from different tables provided I built my own tailored API
