### 2019-01-04 #274 @pachamaltese
@maelle Hi! sorry, I totally forgot to mention that the ROpenSci badge that I added should get another numeral provided 217 is for the older package
