### 2019-01-05 #274 @maelle
### Editor checks:

---

#### Editor comments

Thanks for your submission! I'm now looking for reviewers. Below are a few comments/questions.

In the README please write GPL3 licence to be more specific.

Is it necessary for the repo to be a fork? Or can you at least add an issue tracker to it?

Could you fix the example in `get_countrycode()` docs or at least add another working example?

Also note that when transferring it to the ropensci GitHub organization, it'll need to have the same name as the package.

Please update the review badge, the number in it should be the issue number, 274.

Why is there a CRAN link in DESCRIPTION given the package has not been released on CRAN? By the way please do not submit the package to CRAN before the end of the review process.

There's no example after the sentence "Here’s a basic example to obtain bilateral trade between Chile and Argentina in the year 2000 under different trade classifications:" in the README.

Here's the output from `goodpractice`, for you and the reviewers to have in mind.

```
 ✖ avoid long code lines, it is bad for readability.
    Also, many people prefer editor windows that are about 80
    characters wide. Try make your lines shorter than 80 characters

    R/get_countrycode.R:60:1
    R/get_countrycode.R:67:1
    R/get_data.R:44:1
    R/get_data.R:47:1
    R/get_data.R:97:1
    ... and 15 more lines

  ✖ fix this R CMD check NOTE: Note: found 4 marked UTF-8
    strings
  ✖ fix this R CMD check WARNING: LaTeX errors when
    creating PDF version. This typically indicates Rd problems.
  ✖ fix this R CMD check ERROR: Re-running with no
    redirection of stdout/stderr. Hmm ... looks like a package You
    may want to clean up by 'rm -rf
    /tmp/RtmpYe7MiN/Rd2pdf33297bad420f'
```
Typos

```
cointain      basic-usage.Rmd:45
mesured       basic-usage.Rmd:114
perfoms       get_data.Rd:39
```
---

Reviewers:
Due date:
