### 2019-01-06 #274 @pachamaltese
Hi @maelle
I tested on another laptop and I still cannot reproduce the UTF-8 warnings
Here's my output
```
> pkgdown::build_site()
══ Building pkgdown site ════════════════════════════════════════════════════════════════════════════════════
Reading from: '/home/pacha/github/tradestatistics'
Writing to:   '/home/pacha/github/tradestatistics/docs'
── Initialising site ────────────────────────────────────────────────────────────────────────────────────────
── Building home ────────────────────────────────────────────────────────────────────────────────────────────
Reading 'CODE_OF_CONDUCT.md'
Reading 'README.Rmd'
Writing 'index.html'
── Building function reference ──────────────────────────────────────────────────────────────────────────────
Updating tradestatistics documentation
Writing NAMESPACE
Loading tradestatistics
Writing NAMESPACE
Loading tradestatistics
Reading 'man/countries.Rd'
Reading 'man/get_countrycode.Rd'
Reading 'man/get_data.Rd'
Reading 'man/get_productcode.Rd'
Reading 'man/products.Rd'
Reading 'man/tables.Rd'
── Building articles ────────────────────────────────────────────────────────────────────────────────────────
Reading 'vignettes/basic-usage.Rmd'
Writing 'articles/basic-usage.html'
Reading 'vignettes/creating-datasets.Rmd'
Writing 'articles/creating-datasets.html'
══ DONE ═════════════════════════════════════════════════════════════════════════════════════════════════════
── Previewing site ──────────────────────────────────────────────────────────────────────────────────────────
> devtools::check()
Updating tradestatistics documentation
Writing NAMESPACE
Loading tradestatistics
Writing NAMESPACE
── Building ────────────────────────────────────────────────────────────────────────────── tradestatistics ──
Setting env vars:
● CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
● CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
● CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
─────────────────────────────────────────────────────────────────────────────────────────────────────────────
✔  checking for file ‘/home/pacha/github/tradestatistics/DESCRIPTION’ ...
─  preparing ‘tradestatistics’:
✔  creating vignettes (16.6s)
   Removed empty directory ‘tradestatistics/data-raw’
   Removed empty directory ‘tradestatistics/docs’
─  looking to see if a ‘data/datalist’ file should be added
─  building ‘tradestatistics_0.1.tar.gz’

── Checking ────────────────────────────────────────────────────────────────────────────── tradestatistics ──
Setting env vars:
● _R_CHECK_CRAN_INCOMING_USE_ASPELL_: TRUE
● _R_CHECK_CRAN_INCOMING_REMOTE_    : FALSE
● _R_CHECK_CRAN_INCOMING_           : FALSE
● _R_CHECK_FORCE_SUGGESTS_          : FALSE
── R CMD check ──────────────────────────────────────────────────────────────────────────────────────────────
─  using log directory ‘/tmp/RtmpoNTgn8/tradestatistics.Rcheck’
─  using R version 3.5.2 (2018-12-20)
─  using platform: x86_64-pc-linux-gnu (64-bit)
─  using session charset: UTF-8
─  using options ‘--no-manual --as-cran’
✔  checking for file ‘tradestatistics/DESCRIPTION’
─  checking extension type ... Package
─  this is package ‘tradestatistics’ version ‘0.1’
─  package encoding: UTF-8
✔  checking package namespace information ...
✔  checking package dependencies (1.3s)
✔  checking if this is a source package
✔  checking if there is a namespace
✔  checking for executable files ...
✔  checking for hidden files and directories
✔  checking for portable file names
✔  checking for sufficient/correct file permissions ...
✔  checking serialization versions
✔  checking whether package ‘tradestatistics’ can be installed (1.4s)
✔  checking installed package size ...
✔  checking package directory ...
✔  checking ‘build’ directory
✔  checking top-level files
✔  checking for left-over files
✔  checking index information ...
✔  checking package subdirectories ...
✔  checking R files for non-ASCII characters ...
✔  checking R files for syntax errors ...
✔  checking whether the package can be loaded (394ms)
✔  checking whether the package can be loaded with stated dependencies (358ms)
✔  checking whether the package can be unloaded cleanly (370ms)
✔  checking whether the namespace can be loaded with stated dependencies (361ms)
✔  checking whether the namespace can be unloaded cleanly (388ms)
✔  checking loading without being on the library search path (437ms)
✔  checking dependencies in R code (395ms)
✔  checking S3 generic/method consistency (870ms)
✔  checking replacement functions (385ms)
✔  checking foreign function calls (394ms)
✔  checking R code for possible problems (2.3s)
✔  checking Rd files ...
✔  checking Rd metadata ...
✔  checking Rd line widths ...
✔  checking Rd cross-references ...
✔  checking for missing documentation entries (407ms)
✔  checking for code/documentation mismatches (1.2s)
✔  checking Rd \usage sections (956ms)
✔  checking Rd contents ...
✔  checking for unstated dependencies in examples ...
✔  checking contents of ‘data’ directory
✔  checking data for non-ASCII characters ...
✔  checking data for ASCII and uncompressed saves ...
✔  checking installed files from ‘inst/doc’ ...
✔  checking files in ‘vignettes’ ...
✔  checking examples (976ms)
✔  checking for unstated dependencies in ‘tests’ ...
─  checking tests ...
✔  Running ‘testthat.R’
✔  checking for unstated dependencies in vignettes ...
✔  checking package vignettes in ‘inst/doc’ ...
✔  checking re-building of vignette outputs (30.6s)

── R CMD check results ───────────────────────────────────────────────────────────── tradestatistics 0.1 ────
Duration: 48.7s

0 errors ✔ | 0 warnings ✔ | 0 notes ✔
```
