### 2019-01-07 #274 @maelle
@pachamaltese I've just re-run R CMD Check but not via `devtools::check()` and the warning disappeared :open_mouth: so it was most probably a false positive, sorry about that.
