### 2019-01-07 #274 @maelle
@emilyriederer @mpadge thanks for accepting to review this package! :smiley_cat:  Your reviews are due on 2019-01-28. As a reminder here are our [reviewer guide](https://ropensci.github.io/dev_guide/reviewerguide.html) and [review template](https://ropensci.github.io/dev_guide/reviewtemplate.html).
