### 2019-01-16 #274 @mpadge
- There are currently no contribution guidelines, nor a "CONTRIBUTING" file. As noted by @emilyriederer, the "BugReports" field in the "DESCRIPTION" needs updating.

##### For packages co-submitting to JOSS

- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)

---

Much of the current review does little more than reiterate the phenomenally detailed review of @emilyriederer, to whom I am also indebted for saving me much of the work, and for giving me a very helpful overview of the package!

One of my motivations in reviewing this pacakge was to gain hands-on insight into how @pachamaltese had overcome the server-side problems that plagued the [previous review process](https://github.com/ropensci/software-review/issues/217). The solution is amazingly impressive, and the documentation provided on the [`tradestatistics.io` website](https://tradestatistics.io) goes miles beyond what this package alone might suggest. That is not a problem at all, but some of my suggestions below, and many of the concerns of @emilyriederer, could be very easily allayed by providing links in the documentation to relevant pages of the fantastic [tradestatistics.github.io/documentation](https://tradestatistics.github.io/documentation).

(An immediate example is the central table of [data nomenclature and availability](https://tradestatistics.github.io/documentation/data-availability.html) which is so clear and simple that it could be helpfully reproduced in the documentation for the R package.)

One other general comment, again and as always as already indicated by @emilyriederer, concerns the names of data sets and functions, which it seems fair to suggest are too general, and would be better with, for example, package-specific prefixes (see below). In this regard, I note that the whole [`tradestatistics.io`](https://tradestatistics.io) endeavour seems already set up to be -- or at least to deserve to become -- a mini-empire and go-to place for this stuff. And yet the Observatory of Economic Complexity is not the only source of such data, with arguably an equally rich source being the [OECD](http://www.oecd.org/trade/). I for one could definitely see a case for [`tradestatistics.io`](https://tradestatistics.io) offering a unified interface for accessing such data in general. Please note that I am no way suggesting that ought to be incorporated within the package at the current point in time, and certainly not as any requirement of this review process! I merely mention it as worth thinking about in terms of current nomenclature, because it offers a number of possibilities including,

1. Prefix data and functions with, for example, `oec_get_data()`, and so allow future extension to, for example, `oecd_get_data()`.
2. Provide a general prefix and general functions which sources specified by parameters, `tradestats_get_data(..., source = "oecd")` (or whatever prefix might be desired).

## General style

To continue on the above point, the names of functions and internal data sets are arguably somewhat overly generic, as is much of the function documentation.  @pachamaltese would be advised to keep in mind that this package may be loaded in a workspace with many other packages, and `?counties` simply tells me "Provides details about countries". This should be contextualised to be helpful - "Provides details about countries included in the Observatory of Economic Complexity data retreived by tradestatistics ..." or similar. Analagous arguments apply to most documentation. These are trivial but important fixes.

Function names suffer similar lack of specificity, as already emphasised by @emilyriederer. For example, `rdocumentation.org` suggests 109 functions named `get_data()` in various CRAN packages. Examples of package specific prefixes are [`sf`](https://cran.r-project.org/package=sf) with it's uniform `st_*` functions, or [`piggyback`](https://github.com/ropensci/piggyback) with `pb_*`.  Note, however, that `ts_` is used for the [`tsbox`](https://github.com/christophsax/tsbox) package, so likely best to avoid that.

While I concur with the sentiments of @emilyriederer that the diagram at the outset of the README is uniquely cool, I note that most of the text is a pasted GPL-3 license which is bundled with R and so strictly not needed there.  Github repos are a separate issue, and [Best Practice](https://github.com/coreinfrastructure/best-practices-badge) currently states,

> The project MUST post the license(s) of its results in a standard location in their source repository ... E.g., as a top-level file named LICENSE or COPYING.  License filenames MAY be followed by an extension such as ".txt" or ".md". Note that this criterion is only a requirement on the source repository. You do NOT need to include the license file when generating something from the source code (such as an executable, package, or container). For example, when generating an R package for the Comprehensive R Archive Network (CRAN), follow standard CRAN practice: if the license is a standard license, use the standard short license specification (to avoid installing yet another copy of the text) and list the LICENSE file in an exclusion file such as .Rbuildignore.

Best Practice is nevertheless an optional system, and it would suffice simply to delete the GPL-3 text from README in order to free it up to contain the useful package-related stuff everybody will expect.

## get_countrycode()

While noting @emilyriederer's comment regarding default `NULL` values, I actually think these are generally much safer than having no default values and constantly checking `if (!missing (x))`, because they allow moving straight on to simple type checks for the parameters. I nevertheless note that the following is not very informative:
get_countrycode()
#> Error in if (countryname == "all") { : argument is of length zero
```
In her words, the package "does not explain how NULL values are treated." Moreover, the following call actually returns an object:
get_countrycode(1)
#> There is more than one match for your search. Please try again
#>        using one of these codes:
#> # A tibble: 0 x 6
#> # … with 6 variables: country_iso <chr>, country_name_english <chr>,
#> # country_fullname_english <chr>, continent_id <int>,
#> #   continent <chr>, eu28_member <int>
```
This is because `str_to_lower` returns `countyname = "1"` and the subsequent `str_detect` call then matches that on to all countries which have changed names because these have years of change listed like,
tradestatistics::countries$country_fullname_english [233]
#> [1] "Belgium (Belgium-Luxembourg Customs Union until 1999)"
```
(See comment below for why the result shown on screen is then actually empty.) There are lots of possibilities for handling such cases, and I'm sure @pachamaltese is capable of implementing any sensible option, but I mention this as an incentive to implement a bit more thorough pre-checking of parameter types and classes. (At least ensure argument is stripped to alpha characters only and error on anything else such as numeric.)

Another example:
get_countrycode("par")
#> There is more than one match for your search. Please try again
#>        using one of these codes:
#> # A tibble: 1 x 6
#>   country_iso country_name_english country_fullname_english continent_id continent eu28_member
#>     <chr>       <chr>                <chr>                           <int> <chr> <int>
#>   1 pry         Paraguay             Paraguay                            5   Americas            0
```
That message is misleading because there is actually only one match, contrary to the message. As with the above mismatches of "1" onto years of name changes, this is also a mismatch of "par" onto
  country_iso country_name_english country_fullname_english continent_id continent eu28_member
  <chr>       <chr>                <chr>    <int> <chr>           <int>
1 pry         Paraguay             Paraguay Americas            0
2 all         All Countries        All Countries (this is an alias to show several reporters/partner…
                                                  NA NA                 NA
```
.. and "par" matches on to the protracted and irrelevant "several reporters/partners ..." detail for "All Countries". The "All Countries" is then filtered from the screen dump because that filters `country_name_english` and not `country_fullname_english`, and that is certainly an inconsistency that ought to be rectified!

Also note that there will likely be workflows desiring multiple matches. This currently works, which is good:
get_countrycode("par|arg|vene")
```
but still issues the message, which is undesirable if I actually want all of those. That will be important to keep in mind when converting from `tidyverse` to base R string and index matching. An important starting point may be to presume that any values entered for `countryname` are likely to be the first letters, and so you can at least `grep` on string beginnings, which would then remove both the match of "par" to "...several reporters/partners" and "1" to the guff about name changes. But you'll likely know better than me there, and maybe other alternatives such as partial or fuzzy matching might be preferable?

And because we're here, I repeat the concerns of @emilyriederer, via here example for `get_productcode()`, that the current code for this step uses the potentially unnecessary `tidyverse` functions:
countrycode <- tradestatistics::countries %>%
    dplyr::filter(
           stringr::str_detect(
                      stringr::str_to_lower(!!rlang::sym("country_fullname_english")), countryname
           )
           ) %>%
    dplyr::select(!!rlang::sym("country_iso")) %>%
    purrr::as_vector()
```
while a base R equivalent in this case is actually much clearer and simpler:
index <- grep (countryname,
               tolower (tradestatistics::countries$country_fullname_english))
tradestatistics::countries$country_iso [index]
```

Finally, I suspect issuing a `message` rather than an `error` on `length(countrycode) != 1` is an intentional design feature to prevent the latter from breaking long tidy-type operations, but would perhaps encourage @pachamaltese to consider having `errors` as these can easily be caught in a non-breaking way (for example with the [`collateral` package](https://github.com/rensa/collateral)), yet provide additional important information in such workflows. Happy to leave to authorial judgement call, but I would otherwise suggest that an inability to return a meaningful result ought to be an error.

## get_productcode

I note some potential confusion in the overlap between the `products` data set, the return object of `get_productcode()`, and that of `get_coutrycode()`. The latter does what it says and returns a single code, while `get_productcode()` returns a filtered version of `products` with the search string appended as an extra column. This function thus doesn't really do what it says, rather it returns the full product **table** matching the search item. I'd suggest here that `get_products()` would be a more appropriate name. Maybe @pachamaltese might then see a use case for a `get_productcode()` function that returned just the "commodity_code" column, although some clarification of nomenclature would then be in order as "products" are actually referred to as "commodities". This confusion between "products" and "commodities" pervades the entire package somewhat, and is likely something worth clarifying wherever possible.

I'd also see a use for string matching on `group_name` as well as `product_fullname_english`. Maybe `get_product(productname = NULL, productgroup = NULL)`?

I also don't see any specific need to append the column with the search term to the return result, although grant that may be just a reflection of my ignorance here? Nevertheless, the presence of "group_name" seems to provide an "official" classification of "type_product", and so render this redundant.

## get_data

My main concern here is precisely that of @emilyriederer, and worth quoting:

> What is the envisioned use case for get_productcode() since this information
> can't be passed into get_data()? From this package and the API documentation,
> I gather there is no support for filtering on commodity codes. Is that true? I
> bet a lot of users would be interested in filtering down to a specific
> industry.

My even stronger assertion would be that there would be very few users whose workflow would **not** be `get_data() %>% filter()`. It would be trivial to automate that by implementing the two extra arguments mentioned above,
get_data (..., product_fullname_english = NULL, group_name = NULL)
```
but then if that's to be done, the logical and full extension may as well be implemented,
get_data (..., product_fullname_english = NULL, commodity_code = NULL,
          group_code = NULL, group_name = NULL)
```
(Because you'll likely be using base R rather than `tidyverse` for the filtering, the parameter names need not necessarily be identical with those in the table, but diverging from that may cause confusion.)

To again paraphrase @emilyriederer, in the absence of such filtering ability, it's hard to argue the value of `get_productcode()` (-> `get_products()`?), because it can only be used through the user writing their own code to extract the desired entries, or likely even just as a visual lookup table, in which case it is no different from simply looking at the website.

## Miscellaneous

`DESCRIPTION` has `Suggests: covr`, but this should be removed - it is directly loaded in the `.travis.yml`, and not related to the package itself.

## Checklist

To aid @pachamaltese, this condenses (i think) most of the above points into a simpler checklist. Items with (O) are those I am happy to consider optional and leave the ultimate decision to @pachamaltese.

- [ ] Contributing guidelines; fix DESCR (including removing `covr`)
- [ ] Make function documentation more specific
- [ ] Make names of all functions and data sets more package-specific
- [ ] Remove License text from README (and possibly place elsehwhere), and provide a bit more package-related useful stuff.
- [ ] Fix fail cases and misleading errors for `get_countrycode()`
- [ ] (O) Convert dependency-heavy "tidy" code to base R
- [ ] `get_productcode()` -> `get_product(productname = NULL, productgroup = NULL)` or something along those lines.
- [ ] `get_data()` -> get_data (..., product_fullname_english = NULL, commodity_code = NULL, group_code = NULL, group_name = NULL)` or something along those lines
- [ ] Improve string matching in `get_countycode()`
- [ ] (O) Convert messages on `get_countycode()` failure to errors?

Thanks to @maelle and @pachamaltese for the opportunity to review this package, which has been a great pleasure! It is obviously the result of a phenomenal amount of work culminating in [`tradestatistics.io`](https://tradestatistics.io), of which this R package represents a small component. Congrats @pachamaltese on devising and implementing such an important and consummately well-executed project!
