### 2019-01-16 #274 @pachamaltese
Thanks a lot @maelle , @emilyriederer  and @mpadge
Provided tradestatistics.io was basically an annoyed version of me trying to create a stable and open server to download trade data, I shall start doing some changes to the API so that data filtering by commodity is made in a more efficient way and without downloading data that then you will discard.

Does it work? What do you think about adding a commodity group filtering option to the API and then completing that nice to-do list from above?
