### 2019-01-16 #274 @pachamaltese
Hi @mpadge

Once again, thanks a lot for the useful list.
To achieve some parts of the list without passing unwanted data I had to edit the API during lunch time.
Check this: https://api.tradestatistics.io/yrpc?y=2000&r=chl&p=arg&g=95
`g` is a new parameter that can be any HS group code (i.e. 01, 02, ..., 95, ... 99)
In the example above, the background writes a query like `... WHERE LEFT(commodity_code, 2) = 95`

Another new filtering option in the API is that any table with commodity column (the `g` example also applies to any table with commodity field), for example: https://api.tradestatistics.io/yc?y=1962&c=0101

With these changes I can start improving the package
