### 2019-01-16 #274 @mpadge
That's fantastic, @pachamaltese, and obviously a way more efficient solution than post-filtering. This review is become more and more interesting, with your ability to make server-side changes and then update R code. And all that "during lunch time" :plate_with_cutlery:
