### 2019-01-17 #274 @pachamaltese
@emilyriederer Yes, he API is totally mine. I am not using the OEC (I worked with them for ~3 yrs and they never credited me...but dirty clothes are washed at home!), That's what I am using UN data + my own cleaning scripts.

@emilyriederer @mdpage What do you think about adding a continent alias to the API (e.g. `yrp?y=2000&r=m-asia` where `m-asia` means "all ISO codes of countries in Asia)
