### 2019-01-18 #274 @emilyriederer
@pachamaltese That sounds awesome! I can envision that adding a lot of value for users to have an intermediate step between country-level of global data. The only think I think of these is potential confusion if the continent refers to the reporter, partner, or both.
