### 2019-01-18 #274 @pachamaltese
@emilyriederer what do you think about some "meta-variables" such as `c-am`, `c-af`, etc (i.e. "continent - americas"
