### 2019-01-20 #274 @emilyriederer
@pachamaltese Apologies on my last comment. I realize now I misread yours! I thought the content aliases were being filtered by a new variable and not being passed through `r` and `p`. So please disregard that I said before. Really do love the idea of that sort of flexibility, though, and the `c-{}` structure seems pretty intuitive to me!
