### 2019-02-03 #274 @pachamaltese
@emilyriederer @mpadge
Hi. Sorry for the very long silence.
Today I worked on the changes. Are you ok about splitting the "get data" function in two functions?
I re-organized the data downloading part in:
1. `ots_read_from_api()`
2. `ots_create_tidy_data()`
Being the first more like an internal function, but still easy to use.
