### 2019-02-04 #274 @mpadge
Hi @pachamaltese no worries, and equal apologies my side for not responding to your last couple of questions! As for your current Q: It seems to me that `ots_create_tidy_data()` returns everything that `ots_read_from_api()` does and more, and therefore the latter is not needed at all, I'd suggest. A single function is always less confusing - see analogous discussions in the recent review of [`stats19`](https://github.com/ropensci/software-review/issues/266) [here](https://github.com/ropensci/software-review/issues/266#issuecomment-445504007) and [here](https://github.com/ropensci/software-review/issues/266#issuecomment-449010809) regarding "merge the formatting and reading step". That package also started off with 2 functions, one for downloading/returning the "raw" data, and a second one for processing/re-formatting that data to return a "tidy" equivalent. I agree with the suggestion of @daranzolin there that a single function is simply way less confusing. `stats19` ended up with a single `get_stats19()` function, and I'd suggest a similar thing here: simply revert to a single `ots_read_data()` function by directly renaming `ots_create_tidy_data()`, and removing `ots_read_from_api()` as an exported function.
