### 2019-02-10 #274 @emilyriederer
@pachamaltese Sorry for the slow reply!

@mpadge Are you recommending having fewer total functions or fewer *external* functions? I think the modularization of this function might be related to parts of my feedback that breaking out more internal helper functions might allow for more specific testing. The discussion you point to is very interesting, though, so I can definitely see pros/cons either way. Sorry, @pachamaltese if this feedback was somewhat erroneous!
