### 2019-02-12 #274 @pachamaltese
@mpadge @emilyriederer
sorry for the slow response, in a move to make everyone happy I moved the "read from api part" to an internal function, so that the code is divided in smaller parts and the user is not exposed to a ton of functions

with this I can say all the indications, except for the "less tidyverse part", are complete (I love the tidyverse)
