### 2019-02-12 #274 @mpadge
Sounds great @pachamaltese. And with due apologies for delayed reponse @emilyriederer, I was definitely intending something along the lines of "fewer *external* functions", not in a way necessarily related to testing at all. So pretty much what @pachamaltese has now implemented. And finally does "all the indications ... are complete" suggest you now think the package is sufficiently stable for us to take a more thorough second look?
