### 2019-02-12 #274 @pachamaltese
@mpadge yes! I've edited a lot, you can get a 2nd look and after that I can do more changes if required
