### 2019-03-03 #274 @emilyriederer
Hi @pachamaltese - I haven't read through the code yet, but I notice there are still a few typos in the documentation I might recommend clearing up:

- issues link still broken (change in Description and re-run `pkgdown::build_home()`)
- still missing top-level package documentation (can create with `usethis::use_package_doc()`
- "How to use this package" vignette
    + still calls package `oec`
    + contains broken image reference
- Personally, I think I would still recommend saving the "creating datasets" information just in an R script versus a vignette. It could confuse end-users since this is not something they should ever need to do. Totally fine if you disagree, of course, but in case this comment got lost in my massive review before, I thought I would reshare for consideration
