### 2019-03-04 #274 @pachamaltese
thanks @emilyriederer !! I fixed all of that and with respect to the last comment, I added "This vignette contains something end users should ever need to do"
