### 2019-03-05 #274 @mpadge
With due apologies to @pachamaltese and @maelle for this having taken a bit longer than I suggest above ...

### Second Review

The first thing to note is the now excellent message issued on package load, clearly explaining the terms and limitations of the data license - very happy with that! Neither `devtools::check()` nor `devools::spell_check()` raise any issues at all, so all clear there too. The `ots_` prefixes prepended to all function names are also a welcome improvement, Function documentation has also generally been considerably improved.

### get_countrycode()

The following still need to be fixed:
```{r}
ots_country_code()
#> Error in ots_country_code() : is.character(countryname) is not TRUE
ots_country_code(1)
#> Error in ots_country_code() : is.character(countryname) is not TRUE
```
These errors are still not very informative, and would be better with a more human-centred message like "please provide a country name as a character string".

It might also be useful to include in function examples some examples of actual output from `?ots_country_code` which currently simply has:
```{r}
#' # Single match with no replacement
#' ots_country_code("Chile")
#'
#' # Single match with replacement
#' ots_country_code("America")
#'
#' # Double match with no replacement
#' ots_country_code("Germany")
```
The output of these calls is **not** rendered in the function help, so it would be better to have something more informative like:
```{r}
#' # Single match with no replacement
#' ots_country_code("Chile") # Returns "chl"
#'
#' # Single match with replacement
#' ots_country_code("America") # Returns "usa"
#'
#' # Double match with no replacement
#' ots_country_code("Germany") # Multiple matches; returns "ddr" and "deu"
```

I also note a non-namespaced use of `dplyr::filter` in the code:
```{r}
    f <- tradestatistics::ots_attributes_countries %>%
      filter(
        stringr::str_detect(
          stringr::str_to_lower(!!sym("country_name_english")),
          stringr::str_to_lower(countryname)
        )
      )
```
Should be `dplyr::filter`, not `filter`.

### products
The former lack of clarity between what are now named `ots_attributes_products` (a data set) and `ots_product_code` (a function) has largely been satisfactorily resolved, with the exception of the remaining confusion in nomenclature between product and commodity. Both the table and the function actually contain/return "commodity_code" values, not *product* codes - this confusion really ought to be resolved somehow, either by naming all columns in the table "product_code" and "product_fullname_english" or, alternatively, leaving the tables as they are and renaming both the data and function to "ots_attributes_commodities" and "ots_commodity_code".

My only further note is that is would be useful to have the function avoid the following undesirable behaviour:
```{r}
ots_product_code()
#> Error in ots_product_code() : object 'd' not found
```
I would suggest that it should in this case simply return the whole table. This could be done with a simple line at the outset:
```{r}
if (is.null(productname))
    return (tradestatistics::ots_attributes_products)
```

### ots_create_tidy_data

I merely reiterate my previous comment (with updated function name) that,

> there would be very few users whose workflow would **not** be `ots_create_tidy_data() %>% filter()`. It would be trivial to automate that by implementing ... extra arguments

In your updated and improved form, this still translates to a suggestion to incorporate additional "commodity_code" and "commodity_fullname_english" (presuming renaming as suggested above), because, as already stated, it is otherwise hard to see any convincing reason to have the `ots_attributes_products` table and corresponding `ots_product_codes()` function. These are "helper" functions to focus queries on desired products/commodities, yet there is no direct way in the current implementation to subsequently use the help gained by these functions to directly extract the data.

The documentation Description for `ots_create_tidy_data` has "... and performs different API calls to return tidy data. and data transforming." - that should be fixed.

### Misc

I note that the documentation for the `ots_attributes_` data sets appears to have been manually entered (in `man/ots_attributes_...`) rather than auto-generated by `roxygen2`. This is just a suggestion, and perhaps @pachamaltese has a good reason to keep things that way, but just in case it helps: These can be auto-generated by simply adding the entries in an appropriate `./R` file (`R/tradestatistics-package.R` would seem a good place):
```{r}
#' ots_attributes_countries
#'
#' ...
#' @name ots_attributes_countries
#' @docType data
NULL
```
This does have the distinct advantage of putting the source for all documentation together in the one place - within the `./R` folder.

### checklist (#2)

Again, for your aid @pachamaltese:

- [ ] better examples in `?ots_country_code`
- [ ] More informative error messages for `ots_country_code()`
- [ ] Remove confusion between "product" and "commodity"
- [ ] Have `ots_product_code()` return the full table rather than error
- [ ] Implement additional `commodity/product_code` and `commodity/product_fullname_english` parameters in `ots_create_tidy_data` to implement filtering with single extra parameter.

Hope that helps! Thanks @pachamaltese for the great improvements thus far!
