### 2019-03-10 #274 @pachamaltese
- [ ] better examples in `?ots_country_code`
- [x] More informative error messages for `ots_country_code()`
- [x] Remove confusion between "product" and "commodity"
- [ ] Have `ots_product_code()` return the full table rather than error
- [ ] Implement additional `commodity/product_code` and `commodity/product_fullname_english` parameters in ots_create_tidy_data to implement filtering with single extra parameter.
