### 2019-03-10 #274 @pachamaltese
@mpadge Hi what do you mean with "product_fullname_english parameters in ots_create_tidy_data to implement filtering with single extra parameter"?
