### 2019-03-11 #274 @mpadge
I just mean a very slight modification to current code so that instead of users having to do this in current form:
``` r
x <- ots_create_tidy_data (years = 2015, reporter = "idn", partner = "nld") # 1,067 rows
x %>% dplyr::filter (grepl ("spice", x$group_name, ignore.case = TRUE)) # 20 rows
x %>% dplyr::filter (grepl ("ginger", x$product_fullname_english, ignore.case = TRUE)) # 1 row
```
where the `group_name` or `product_fullname_english` have to be manually matched by post-extraction filtering, this could simply be incorporated as an additional line at the end of the code:
``` r
ots_create_tidy_data <- function (..., group_name = NULL, product_fullname_english = NULL) {
...
    if (!is.null(group_name))
        data <- data %>% dplyr::filter (grepl (group_name, data$group_name, ignore.case = TRUE))
    if (!is.null (product_fullname_english))
        data <- data %>% dplyr::filter (grepl (product_fullname_english,
            data$$product_fullname_english, ignore.case = TRUE))
    if (nrow (data) == 0)
        message("No data returned, perhaps try ... [some kind of informative message here]")
    return (data)
}
```
That would then enable the much more intuitive interface
```
x <- ots_create_tidy_data (years = 2015, reporter = "idn", partner = "nld",
        product_fullname_english = "ginger")
```
It is of course up to you whether you might like to think about performing the filtering at the rest API level, because in this example case the amount of data finally required is > 1,000 times less. Either way, motivation is simply my impression that a very general use of these data will always be to implement such post-filtering anyway, so it would make the package more user friendly to implement this simple change.
