### 2019-03-11 #274 @mpadge
Another thought: This is just a suggestion and entirely a matter of personal style, but since you're importing `magrittr` anyway, you could also import `%<>%` and so replace all current lines like
``` r
data <- data %>% some_kind_of_verb(...)
```
with
``` r
data %<>% some_kind_of_verb(...)
```
There are definitely people who would consider the second version somewhat easier to read and understand - myself included. Entirely up to you of course ...
