### 2019-03-11 #274 @pachamaltese
@mpadge I opted for something a bit simpler. What do you think about this?
```r
ots_create_tidy_data(year = 1980:1981, reporter = "chl", partner = "arg", product = "0101")
ots_create_tidy_data(year = 1980:1981, reporter = "chl", partner = "arg", product = "01")
ots_create_tidy_data(year = 1980:1981, reporter = "chl", partner = "arg", product = "apple")
```

the second returns all the products in group 01 while the third returns all the rows that contain "apple" an product and/or group name.
