### 2019-03-11 #274 @mpadge
Yeah, that's a good solution too. My main concern was that without this kind of functionality, there was no direct link between the product/commodity tables and the actual trade volumes. The ability to at least put the product codes in is sufficient to make that direct link, and in my opinion seems likely to make the package much more directly usable by more people. Nice!
