### 2019-03-11 #274 @pachamaltese
@mpadge thanks !! I also renamed all the relevant fields to `product` (and therefore I had to create the compressed filed from tradestatistics.io again as well as updating the code and processing repo)
