### 2019-03-11 #274 @mpadge
Great stuff @pachamaltese - sorry if that was a hassle for you, but i really did think that the terminological confusion of asking for "products" and getting "commodities" was really worth resolving. Thanks!
