### 2019-03-11 #274 @pachamaltese
no worries @mpadge, actually UN Comtrade "commodity" nomenclature is confusing, if you look product tables the HS codes contain much more than commodities
