### 2019-03-14 #274 @pachamaltese
@mpadge ready! I implemented `%<>%`
