### 2019-03-16 #274 @emilyriederer
Apologies for my slow follow-up, and thanks to @mpadge and @pachamaltese for the great reviews / discussions which gave me a lot of great context.

# General

Great job, @pachamaltese ! A lot of the changes you’ve made I think have taken an already very good package to the next level. The level of documentation and the carefully crafted API should make this package quite accessible for new users to learn and use - that is great!

# Package Maintenance

Thanks for adding more documentation / checks around which years of data are valid!

One question on package upkeep: do you plan to add additional years of trade data to your API going forward? If so, it may be awkward for a maintainer to manually change the `all(year %in% 1962:2016) != TRUE` condition in `ots_get_tidy_data()` and to update the documentation / website accordingly.

One potential solution would be to build in an API endpoint that reports out the maximum year of data contained and an internal function in the R package that can dynamically make that API call and use the output in `ots_get_tidy_data()`.

Alternatively, this should at least somehow be well-documented for maintainers that API  changes should lead to multiple changes in the `ots_get_tidy_data()` file and the artifacts it generates.

# Documentation

- Nice changes to the vignette! I felt like the examples were a lot more illuminating / use-case motivated for me which makes it a fun and very useful read
- `ots_get_tidy_data()` has a `year` parameter which can conveniently be more than one year. However, the documentation reads "`Numeric value greater or equal to 1962 and lower of equal to 2016. Default set to NULL`" which I worry users will interpret to mean a single year. Perhaps "value or vector"

# API

- I still wonder if it would make sense to have a consistent output type for `ots_product_code()` and `ots_country_code()`. The former returns a `tibble` regardless of 0, 1, or multiple matches where as the later sometimes returns a vector and other times a country code. Personally, as a user, the inconsistency *between* these two functions and *within* the `ots_country_code()` function is a bit confusing to me. I'm sure there's a good reason for it to be this way, but perhaps this can also somehow be highlighted in the documentation?

# Dependencies

Of course, you should use whichever of base of `tidyverse` that you think is best for you and your users! That said, I might still advocate for swapping out `stringr` functions with the base equivalents I mentioned above since that is a rather large package and you are using so few functions from it. Again though, I know this is totally a personal style thing so not trying to be annoying! 🙂

# Implementation

I love the more helpful / informative error messages with the argument checks! ❤️

# Checklist

- [ ]  Help maintainers keep date range code / documentation in sync with API
- [ ]  Consider rephrasing `year` parameter documentation for `ots_get_tidy_data()`
- [ ]  Consider output types of `ots_country_code()` (sometimes vector sometimes tibble) versus `ots_product_code()` and standardize or document
