### 2019-03-16 #274 @pachamaltese
Hi @emilyriederer !

Thanks a lot !! The only part I'm not sure is dropping `stringr` since I'm using `str_detect`, `str_to_lower`, `str_trim` and `str_squish`. I'm aware that `gsub(..., perl = TRUE)` can replace that but still I'm not sure about readability.

With respect to the other parts:

- [x] *Help maintainers keep date range code / documentation in sync with API.* Ready, I added https://api.tradestatistics.io/year_range and it's used in the package and documentation
- [x] *Consider rephrasing year parameter documentation for ots_get_tidy_data().* Ready, not just that, I also edited the code to allow multiple parameters like `ots_create_tidy_data(years = 1980, reporters = c("chl", "per"), partners = c("arg", "bra"), products = c("0101", "apple"))` that was not possible until you suggested to improve the docs, so I went to the very end.
- [x]  *Consider output types of ots_country_code() (sometimes vector sometimes tibble) versus ots_product_code() and standardize or document.* Ready, with all the editions needed in other parts of the package. Now both product and country searches shall return a tibble.
