### 2019-03-17 #274 @emilyriederer
Thanks @pachamaltese - makes sense! Whatever you think is best 🙂

Thanks for considering my other thoughts. I love these last changes that you've made! It's such a cool and unique thing that you're in charge of both the API and the wrapper package! It's really been interesting opening up the possibility of changing either to improve the experience.  Also *very* cool that not all the functions can take multiple options like that!
