### 2019-03-17 #274 @pachamaltese
Thanks a lot @emilyriederer , and of course also to @mpadge for helping to make this package better. Just blame @jbkunst that 3 yrs ago insisted that I should drop base+data.table and move to the tidyverse :P
