### 2019-03-27 #274 @maelle
Ok! @emilyriederer @mpadge, if you're happy with the changes implemented, can you check the box "The author has responded to my review and made changes to my satisfaction. I recommend approving this package." in your reviews?

Thanks to all of you for a very active and productive discussion in this thread!
