### 2019-04-01 #274 @pachamaltese
@maelle @emilyriederer @mpadge can you help me sharing this dashboard that is basically a GUI for the package? https://shiny.tradestatistics.io/
the main website is https://tradestatistics.io/
