### 2019-04-02 #274 @mpadge
Completed above and copied here for completeness:

#### Final approval (post-review)

I've sent a very minor pull request of the [only outstanding issue](https://github.com/tradestatistics/tradestatistics/pull/1). Thanks so much @pachamaltese for the amazing work behind the scenes with your own server, and for the seamless package interface. I'll definitely be using this package!
