### 2019-04-03 #274 @pachamaltese
I've completed these steps:

- [x] Transfer the repo to rOpenSci's "ropensci" GitHub organization under "Settings" in your repo. I have invited you to a team that should allow you to do so. You'll be made admin again once you do.
- [x] Add the rOpenSci footer to the bottom of your README " [![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)"
- [x] Fix any links in badges for CI and coverage to point to the ropensci URL. We no longer transfer Appveyor projects to ropensci Appveyor account so after transfer of your repo to rOpenSci's "ropensci" - [x] GitHub organization the badge should be [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ropensci/pkgname?branch=master&svg=true)](https://ci.appveyor.com/project/individualaccount/pkgname).

Also important:
- [ ] I also updated all links in DESCRIPTION, Readme and vignettes that point to tradestatistics.github.io/* to ropensci.github.io. Is that ok?
- [ ] Besides it, I kept Joshua, Amanda and Jorge as ctb provided they gave tons of feedback on previous iterations. Shall I keep them?
- [x] UN Comtrade is be credited as dtc.
- [ ] Shall I add Emily and Mark as ctb?
