### 2019-04-03 #274 @maelle
Yes to keeping previous contributors over time, and regarding reviewers, you should ask them :wink:
