### 2019-04-03 #274 @pachamaltese
not yet, I am clearing ctb before pushing again before moving the repository
> I don't see your repo under the ropensci org, are you sure you've transferred it?
