### 2019-04-03 #274 @pachamaltese
@maelle The only thing I cannot change is the repo description in github:
![image](https://user-images.githubusercontent.com/10091065/55490143-a4bc2980-5609-11e9-97a3-ad1737c47c25.png)
It should say https://ropensci.github.io/tradestatistics but I cannot edit that header
