### 2019-04-03 #274 @pachamaltese
I also can't change "R package to access our API" to "R package to access Open Trade Statistics API" since the repo has moved
