### 2019-04-03 #274 @maelle
Speaking of `pkgdown` feel free to use the brand-new rOpenSci style https://ropensci.github.io/rotemplate/ (feedback on the package and its guidance very welcome).
