### 2019-04-04 #274 @maelle
By the way @pachamaltese we recommend the role "rev" for reviewers since it's more specific than "ctb". Folks can also be both, of course.

To easily add folks to DESCRIPTION I recommend e.g. `desc::desc_add_author_gh("emilyriederer", role = "rev")`.
