### 2019-04-04 #274 @maelle
By the way @pachamaltese we recommend the role "rev" for reviewers since it's more specific than "ctb". Folks can also be both, of course.

To easily add folks to DESCRIPTION I recommend e.g. `desc::desc_add_author_gh("emilyriederer", role = "rev", comment = "reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/274"))` (I'm planning on adding a better helper to `rodev` that'd add the comment automatically).
