### 2019-04-04 #274 @pachamaltese
thanks @maelle !!
I was trying to add
```

      person(given = "Mark",
             family = "Padgham",
             role = "rev",
             comment = c("reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/274",
                         ORCID = "0000-0003-2172-5265")),
```
but comment is limited to a single string, probably I can create a PR to ease that
