### 2019-04-04 #274 @maelle
Something like

```r
desc::desc_add_author(given = "Mark",
                      family = "Padgham",
                      role = "rev",
                      comment = "blabla", orcid = "orcidid")
```
works although the comment section looks ugly.
